+++ src/js/language/sv_SE.js
@@ -175,6 +175,7 @@
                 SI: 'Slovenien',
                 SK: 'Slovakien',
                 SM: 'San Marino',
+                TH: 'Thailand',
                 ZA: 'Sydafrika'
             }
         },
@@ -237,6 +238,7 @@
                 MA: 'Marocko',
                 PK: 'Pakistan',
                 RO: 'Rumänien',
+                TH: 'Thailand',
                 US: 'USA'
             }
         },